Split ParameterNamingSpec
